requirements.txt
streamlit
sqlalchemy
python-dotenv
psycopg2